refactor/feat: mover tratamento e adicionar paginação na listagem
