+++ backend/src/main/java/com/soulCodeJr/backEndSite/controllers/CandidateResource.java
@@ -1,13 +1,13 @@
-package com.soulCodeJr.backEndSite.resources;
+package com.soulCodeJr.backEndSite.controllers;
 
-import com.soulCodeJr.backEndSite.dto.ErrorMessage;
 import com.soulCodeJr.backEndSite.entities.Candidate;
 import com.soulCodeJr.backEndSite.services.CandidateService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.web.PageableDefault;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
 
 @RestController
 @RequestMapping(value = "/candidates")
@@ -16,19 +16,17 @@ public class CandidateResource {
     private CandidateService candidateService;
 
     @GetMapping
-    public ResponseEntity<List<Candidate>> findAll() {
-        List<Candidate> candidates = candidateService.findAll();
-        return ResponseEntity.ok().body(candidates);
+    public ResponseEntity<Page<Candidate>> findAll(
+            @PageableDefault(size = 5, sort = "name") Pageable pageable
+    ) {
+        Page<Candidate> candidates = candidateService.findAll(pageable);
+        return ResponseEntity.ok(candidates);
     }
 
     @GetMapping(value = "{id}")
-    public ResponseEntity<?> findById(@PathVariable Long id) {
-        try {
-            Candidate candidate = candidateService.findById(id);
-            return ResponseEntity.ok().body(candidate);
-        } catch (Exception e) {
-            return ResponseEntity.status(404).body(new ErrorMessage(e.getMessage()));
-        }
+    public ResponseEntity<Candidate> findById(@PathVariable Long id) {
+        Candidate candidate = candidateService.findById(id);
+        return ResponseEntity.ok().body(candidate);
     }
 
     @PostMapping("/candidates")
