+++ backend/src/main/java/com/soulCodeJr/backEndSite/exception/ErrorMessage.java
@@ -1,4 +1,4 @@
-package com.soulCodeJr.backEndSite.dto;
+package com.soulCodeJr.backEndSite.exception;
 
 public record ErrorMessage(String message) {
 }
