+++ backend/src/main/java/com/soulCodeJr/backEndSite/exception/UserIndicatorNotFoundException.java
@@ -1,5 +1,6 @@
 package com.soulCodeJr.backEndSite.exception;
 
+//Deixarei para caso seja necessário usar para alguma coisa
 public class UserIndicatorNotFoundException extends RuntimeException {
     public UserIndicatorNotFoundException(String message) {
         super(message);
