+++ backend/src/main/resources/application.properties
@@ -5,5 +5,5 @@ spring.jpa.open-in-view=false
 cors.origins=${CORS_ORIGINS:http://localhost:5173,http://localhost:3000}
 
 #Caso ocorra o problema do SpringSecurity
-#spring.security.user.name=admin
-#spring.security.user.password=1234
+spring.security.user.name=admin
+spring.security.user.password=1234
